9 회원가입 로그인 session 활용한 인증

+++ src/main/java/com/widyu/healthcare/mapper/GuardiansMapper.java
@@ -12,5 +12,5 @@ public interface GuardiansMapper {
     public int checkId(String id);
     public int insert(UsersDTO encryptedUser);
     public UsersDTO findByIdAndPassword(@Param("id") String id, @Param("password") String password);
-    public List<UsersDTO> findSeniorsById(@Param("userIdx") Integer userIdx);
+    public List<UsersDTO> findSeniorsById(@Param("userIdx") long userIdx);
 }
